+++ src/main/java/com/dao/HotelAmenityDAO.java
@@ -12,6 +12,6 @@ import com.model.HotelAmenityId;
 @Repository
 public interface HotelAmenityDAO extends JpaRepository<HotelAmenity, HotelAmenityId> {
 	
-	 List<HotelAmenity> findByAmenity_AmenityId(int amenityId);
+	 List<HotelAmenity> findByAmenity_AmenityId(long amenityId);
 	
-	 List<HotelAmenity> findByHotel_HotelId(int hotelId);}
+	 List<HotelAmenity> findByHotel_HotelId(long hotelId);}
