+++ src/main/java/com/dao/HotelDAO.java
@@ -9,7 +9,7 @@ import com.model.Hotel;
 import com.model.Room;
 
 @Repository
-public interface HotelDAO extends JpaRepository<Hotel, Integer> {
+public interface HotelDAO extends JpaRepository<Hotel, Long> {
 	
     @Query("SELECT r FROM Hotel r WHERE r.location = :location")
     List<Room> findHotelsByLocation(String location);
