+++ src/main/java/com/dao/ReservationDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.model.Reservation;
 @Repository
-public interface ReservationDAO extends JpaRepository<Reservation, Integer> {
+public interface ReservationDAO extends JpaRepository<Reservation, Long> {
 	
 	 List<Reservation> findByCheckInDateBetween(Date startDate, Date endDate);
 }
