+++ src/main/java/com/dao/TaskDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.model.Task;
 
 @Repository
-public interface TaskDAO extends JpaRepository<Task,Integer> {
+public interface TaskDAO extends JpaRepository<Task,Long> {
 
 }
