+++ src/main/java/com/model/Reservation.java
@@ -11,7 +11,7 @@ import jakarta.persistence.OneToOne;
 public class Reservation {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int reservation_id;
+	private long reservation_id;
 	private String guest_name;
 	private String guest_email;
 	private String guest_phone;
@@ -24,11 +24,11 @@ public class Reservation {
 
 
 
-	public int getReservation_id() {
+	public long getReservation_id() {
 		return reservation_id;
 	}
 
-	public void setReservation_id(int reservation_id) {
+	public void setReservation_id(long reservation_id) {
 		this.reservation_id = reservation_id;
 	}
