+++ src/main/java/com/service/PaymentService.java
@@ -20,7 +20,7 @@ public class PaymentService {
     public List<Payment> getAll() {
         return paymentDAO.findAll();
     }
-    public Payment getById(int id) {
+    public Payment getById(long id) {
         return paymentDAO.findById(id).orElse(null);
     }
    public List<Payment> getByStatus(String status) {
@@ -31,7 +31,7 @@ public class PaymentService {
         return payments.stream().mapToDouble(Payment::getAmount).sum();
     }
 
-    public boolean delete(int id) {
+    public boolean delete(long id) {
         Payment payment = paymentDAO.findById(id).orElse(null);
         if (payment != null) {
             paymentDAO.delete(payment);
